% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{.build_formula}
\alias{.build_formula}
\title{Build an lm() formula from components}
\usage{
.build_formula(dep, contset, fes_label)
}
\arguments{
\item{dep}{Character scalar (dependent variable).}

\item{contset}{Character scalar like "CONTa+CONTb".}

\item{fes_label}{Character scalar FE label (see \code{.fes_to_formula}).}
}
\value{
A formula object \code{dep ~ ...}.
}
\description{
Build an lm() formula from components
}
\keyword{internal}
