% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{.fes_to_formula}
\alias{.fes_to_formula}
\title{Map fixed-effects label to a formula fragment}
\usage{
.fes_to_formula(fes_label)
}
\arguments{
\item{fes_label}{Character scalar. One of "none","time","group","time+group".}
}
\value{
Character string (RHS fragment) or error on unknown label.
}
\description{
Map fixed-effects label to a formula fragment
}
\keyword{internal}
