% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{plot_importance_mlogit}
\alias{plot_importance_mlogit}
\title{Plot feature importance for mlogit approach}
\usage{
plot_importance_mlogit(out, aggregate = TRUE)
}
\arguments{
\item{out}{Output list from \code{find_uncertainty_source_mlogit()}.}

\item{aggregate}{Logical. If \code{TRUE}, aggregates to dimensions.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Visual summary of average absolute probability changes by dimension or by
specific feature/value when using the MLOGIT method.
}
\seealso{
Other plot-sources: 
\code{\link{plot_importance_neigh}()},
\code{\link{plot_importance_neuronet}()}
}
\concept{plot-sources}
