% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{plot_importance_neuronet}
\alias{plot_importance_neuronet}
\title{Plot importances (SHAP) of neural network approach}
\usage{
plot_importance_neuronet(out, aggregate = TRUE)
}
\arguments{
\item{out}{Output of \code{find_uncertainty_source_neuronet()}.}

\item{aggregate}{Logical. If \code{TRUE}, plot dimension aggregates.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Displays either dimension-level average SHAP shares (\code{aggregate = TRUE})
or feature-level SHAP shares (\code{aggregate = FALSE}), split by outcome class.
}
\seealso{
Other plot-sources: 
\code{\link{plot_importance_mlogit}()},
\code{\link{plot_importance_neigh}()}
}
\concept{plot-sources}
