% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\docType{package}
\name{sensiverse-package}
\alias{sensiverse}
\alias{sensiverse-package}
\title{sensiverse}
\description{
Tools to estimate, filter, and interpret large model spaces for specification
sensitivity. The package provides (i) fast OLS grid estimation, (ii) filters
for curated universes, (iii) sign-share diagnostics with plots, and (iv) three
complementary methods to pinpoint uncertainty sources (MLOGIT, 1NN,
and NN+SHAP).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.pnas.org/doi/10.1073/pnas.2414926122}
  \item Report bugs at \url{https://github.com/MGanslmeier/sensiverse/issues}
}

}
\author{
\strong{Maintainer}: Michael Ganslmeier \email{m.ganslmeier@exeter.ac.uk}

Authors:
\itemize{
  \item Tim Vlandas \email{t.vlandas@spi.ox.ac.uk}
}

}
\keyword{internal}
